pyproject.toml
[project]
name = "video-rag"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    # Core
    "pixeltable>=0.2.26",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "loguru",
    # API
    "fastapi",
    "uvicorn[standard]",
    "python-multipart", # file uploads
    "click",
    # MCP
    "fastmcp",
    # AWS / Nova
    "botocore",
    # Media processing
    "av", # PyAV for video
    "moviepy",
    "pillow",
    "opencv-python-headless",
    # Observability
    "opik",
    # Misc
    "instructor",
    "boto3>=1.42.59",
    "aws-sdk-bedrock-runtime>=0.4.0",
]

[project.scripts]
serve-mcp = "video_rag.mcp.server:run_mcp"
serve-api = "video_rag.api.app:run_api"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/video_rag"]

[tool.uv]
dev-dependencies = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "ruff",
]
